fix: Restaurant 서비스 Outbox OutboxStatus 제거 및 ProcessedMessage 엔티티 추가해 중복 메시지 처리

+++ restaurant/src/main/java/com/orderingsystem/restaurant/application/OwnerService.java
@@ -2,7 +2,10 @@ package com.orderingsystem.restaurant.application;
 
 import com.orderingsystem.restaurant.application.dto.request.RestaurantOwnerApplicationRequest;
 import com.orderingsystem.restaurant.domain.model.Owner;
+import com.orderingsystem.restaurant.domain.model.outbox.MessageType;
 import com.orderingsystem.restaurant.domain.repository.OwnerRepository;
+import com.orderingsystem.restaurant.domain.repository.outbox.ProcessedMessageRepository;
+import java.time.ZonedDateTime;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -14,9 +17,14 @@ import org.springframework.transaction.annotation.Transactional;
 public class OwnerService {
 
     private final OwnerRepository ownerRepository;
+    private final ProcessedMessageRepository processedMessageRepository;
 
     @Transactional
     public void createOwner(RestaurantOwnerApplicationRequest restaurantOwnerApplicationRequest) {
+        if (markIfNotProcessed(restaurantOwnerApplicationRequest, MessageType.OWNER_CREATE)) {
+            return;
+        }
+
         log.info("레스토랑 오너를 생성합니다. Owner Id : {}", restaurantOwnerApplicationRequest.getId());
 
         ownerRepository.save(Owner.builder()
@@ -27,9 +35,28 @@ public class OwnerService {
 
     @Transactional
     public void deleteOwner(RestaurantOwnerApplicationRequest restaurantOwnerApplicationRequest) {
+        if (markIfNotProcessed(restaurantOwnerApplicationRequest, MessageType.OWNER_DELETE)) {
+            return;
+        }
+
         log.info("레스토랑 오너를 삭제합니다. Owner Id : {}", restaurantOwnerApplicationRequest.getId());
 
         ownerRepository.deleteById(restaurantOwnerApplicationRequest.getId());
+    }
+
+    private boolean markIfNotProcessed(RestaurantOwnerApplicationRequest restaurantOwnerApplicationRequest,
+                                       MessageType ownerCreate) {
+        int inserted = processedMessageRepository.insertIgnore(
+                restaurantOwnerApplicationRequest.getId(),
+                ownerCreate.name(),
+                ZonedDateTime.now()
+        );
+
+        if (inserted == 0) {
+            log.info("이미 처리된 Owner {} 메시지입니다. Owner Id : {}", ownerCreate, restaurantOwnerApplicationRequest.getId());
+            return true;
+        }
+        return false;
     }
 
 }
